requirements.txt
azure-functions 
azure-identity 
azure-keyvault-secrets 
azure-storage-blob 
opencensus-ext-azure 
openai 
joblib 
scikit-learn 
flask 
pandas  
python-dotenv

-e .